requirements.txt
# Core dependencies
numpy>=1.19.0
torch>=1.7.0
torchvision>=0.8.0

# Image processing
opencv-python>=4.5.0
Pillow>=8.0.0
scikit-image>=0.18.0
scikit-learn>=0.24.0
scipy>=1.5.0
albumentations>=0.5.2

# Visualization
matplotlib>=3.3.0
tensorboard>=2.4.0
seaborn>=0.11.0

# CRF for post-processing
pydensecrf>=1.0rc3

# Metrics
torchmetrics>=0.6.0

# Configuration
pyyaml>=5.4.0

# Deep learning utilities
tqdm>=4.50.0

# Data manipulation
pandas>=1.3.0

# For machine learning analysis
shap>=0.39.0
imbalanced-learn>=0.8.0
joblib>=1.0.0

# For parallel processing (in analysis.py)
multiprocessing
